+++ src/api/DomainObjects.ts
@@ -234,6 +234,7 @@ export class RequirementsRequest implements IRequirementsRequestCRUD {
 export interface IRequestValidation {
     IsErrored?: boolean,
     TitleError: string,
+    RequestDateError: string,
     RequesterError: string,
     RequesterOrgSymbolError: string,
     RequesterDSNPhoneError: string,
@@ -276,9 +277,22 @@ export class RequestValidation {
         }
     }
 
+    private static getDateValidation(field: Moment, minDate?: Moment, maxDate?: Moment): string {
+        if (!field) {
+            return "Please enter a date that the requirement is needed by!";
+        } else if (minDate && field.isBefore(minDate.startOf('day'))) {
+            return `Your date selected must be on or after ${minDate.format("DD MMM YYYY")}!`;
+        } else if (maxDate && field.isAfter(maxDate.endOf('day'))) {
+            return `Your date selected must be on or before ${maxDate.format("DD MMM YYYY")}!`;
+        } else {
+            return "";
+        }
+    }
+
     static getValidation(request: IRequirementsRequest, isFunded: boolean): IRequestValidation {
         let validation: IRequestValidation = {
             TitleError: this.getSingleLineValidation(request.Title, 255),
+            RequestDateError: this.getDateValidation(request.RequestDate, undefined, moment()),
             RequesterError: request.Requester && request.Requester.EMail ? "" : "Please provide a Requester!",
             RequesterOrgSymbolError: this.getSingleLineValidation(request.RequesterOrgSymbol,15),
             RequesterDSNPhoneError: this.getPhoneNumberValidation(request.RequesterDSNPhone),
@@ -291,7 +305,7 @@ export class RequestValidation {
             OtherApplicationNeededError: request.ApplicationNeeded === ApplicationTypes.OTHER ? this.getSingleLineValidation(request.OtherApplicationNeeded, 255) : "",
             ProjectedOrgsImpactedOrgError: this.getSingleLineValidation(request.ProjectedOrgsImpactedOrg, 15),
             ProjectedImpactedUsersError: request.ProjectedImpactedUsers > 0 ? "" : "Please enter the projected number of users to be impacted by the request!",
-            OperationalNeedDateError: request.OperationalNeedDate ? "" : "Please enter a date that the requirement is needed by!",
+            OperationalNeedDateError: this.getDateValidation(request.OperationalNeedDate, moment()),
             PriorityExplanationError: request.PriorityExplanation ? "" : "Please enter an explanation for why the priority of the requirements request was given!",
             BusinessObjectiveError: request.BusinessObjective ? "" : "Please enter the business objective for the requirement being requested!",
             FunctionalRequirementsError: request.FunctionalRequirements ? "" : "Please enter the functional requirements for the requirement being requested!",